+++ dsmr_frontend/tests/webinterface/test_trends.py
@@ -1,4 +1,5 @@
 from unittest import mock
+import json
 
 from django.test import TestCase, Client
 from django.utils import timezone
@@ -29,28 +30,63 @@ class TestViews(TestCase):
         self.user = User.objects.create_user('testuser', 'unknown@localhost', 'passwd')
         dsmr_consumption.services.compact_all()
 
-    @mock.patch('django.utils.timezone.now')
-    def test_trends(self, now_mock):
-        now_mock.return_value = timezone.make_aware(timezone.datetime(2016, 1, 1))
+    def test_trends(self):
         response = self.client.get(
             reverse('{}:trends'.format(self.namespace))
         )
         self.assertEqual(response.status_code, 200)
         self.assertIn('capabilities', response.context)
-        self.assertIn('day_statistics_count', response.context)
-        self.assertIn('hour_statistics_count', response.context)
+        self.assertIn('frontend_settings', response.context)
 
-        if 'avg_consumption_x' in response.context:
-            self.assertIn('electricity_by_tariff_week', response.context)
-            self.assertIn('electricity_by_tariff_month', response.context)
+    @mock.patch('django.utils.timezone.now')
+    def test_trends_xhr_avg_consumption(self, now_mock):
+        now_mock.return_value = timezone.make_aware(timezone.datetime(2016, 1, 1))
+        response = self.client.get(
+            reverse('{}:trends-xhr-avg-consumption'.format(self.namespace))
+        )
+        self.assertEqual(response.status_code, 200)
+        json_response = json.loads(response.content.decode("utf-8"))
+
+        if not self.support_data:
+            return self.assertEqual(json_response, {"electricity": [], "electricity_returned": [], "gas": []})
+
+        self.assertEqual(len(json_response['electricity']), 24)
+        self.assertEqual(len(json_response['electricity_returned']), 24)
+
+        if self.support_gas:
+            self.assertEqual(len(json_response['gas']), 24)
+        else:
+            self.assertEqual(len(json_response['gas']), 0)
 
         # Test with missing electricity returned.
         ElectricityConsumption.objects.all().update(currently_returned=0)
+
         response = self.client.get(
-            reverse('{}:trends'.format(self.namespace))
+            reverse('{}:trends-xhr-avg-consumption'.format(self.namespace))
+        )
+        json_response = json.loads(response.content.decode("utf-8"))
+
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(len(json_response['electricity_returned']), 0)
+
+    @mock.patch('django.utils.timezone.now')
+    def test_trends_xhr_by_tariff(self, now_mock):
+        now_mock.return_value = timezone.make_aware(timezone.datetime(2016, 1, 1))
+        response = self.client.get(
+            reverse('{}:trends-xhr-consumption-by-tariff'.format(self.namespace))
         )
         self.assertEqual(response.status_code, 200)
-        self.assertFalse(response.context['capabilities']['electricity_returned'])
+        json_response = json.loads(response.content.decode("utf-8"))
+
+        if not self.support_data:
+            return self.assertEqual(json_response, {})
+
+        self.assertIn('week', json_response)
+        self.assertIn('month', json_response)
+        self.assertIn({'value': 84, 'name': 'electricity1'}, json_response['week'])
+        self.assertIn({'value': 16, 'name': 'electricity2'}, json_response['week'])
+        self.assertIn({'value': 84, 'name': 'electricity1'}, json_response['month'])
+        self.assertIn({'value': 16, 'name': 'electricity2'}, json_response['month'])
 
 
 class TestViewsWithoutData(TestViews):